// src/app/core/Interfaces/interfaces.ts
export interface Users {
  id: string;
  email: string;
  password: string;
  username: string;
  place: string;
  designation: string;
  birthdate: string;
  country: string;
  facebook: string;
  instagram: string;
  linkedIn: string;
  profileimage:string;
}


export interface tasks {
    id: any;
    title: string;
    description: string;
    dueDate: string;
    priority: string;
    status: string;
  }

  